Revert "enh: Rename private sources, include public headers with rel path"
