+++ src/gflags_completions.cc
@@ -46,6 +46,11 @@
 //     5a) Force bash to place most-relevent groups at the top of the list
 //     5b) Trim most flag's descriptions to fit on a single terminal line
 
+
+#include "gflags_completions.h"
+
+#include "config.h"
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>   // for strlen
@@ -55,11 +60,8 @@
 #include <utility>
 #include <vector>
 
-#include "config.h"
+#include "gflags.h"
 #include "util.h"
-
-#include "gflags/gflags.h"
-#include "gflags/gflags_completions.h"
 
 using std::set;
 using std::string;
